号段查看 异常

+++ icop/src/main/java/com/bank/icop/vo/VoucherNumberVo.java
@@ -19,6 +19,6 @@ public class VoucherNumberVo implements Serializable {
     private String endNo;
 
     @ApiModelProperty(value = "数量  ")
-    private Integer num;
+    private long num;
 
 }
